+++ MaxMind.MinFraud.UnitTest/Request/DeviceTest.cs
@@ -7,20 +7,19 @@ namespace MaxMind.MinFraud.UnitTest.Request
 {
     public class DeviceTest
     {
-        private IPAddress IP { get; } = IPAddress.Parse("1.1.1.1");
-
         [Fact]
         public void TestIPAddress()
         {
-            var device = new Device(ipAddress: IP);
-            Assert.Equal(IP, device.IPAddress);
+            var ip = IPAddress.Parse("1.1.1.1");
+            var device = new Device(ipAddress: ip);
+            Assert.Equal(ip, device.IPAddress);
         }
 
         [Fact]
         public void TestUserAgent()
         {
             var ua = "Mozila 5";
-            var device = new Device(ipAddress: IP, userAgent: ua);
+            var device = new Device(userAgent: ua);
             Assert.Equal(ua, device.UserAgent);
         }
 
@@ -28,34 +27,34 @@ namespace MaxMind.MinFraud.UnitTest.Request
         public void TestAcceptLanguage()
         {
             var al = "en-US";
-            var device = new Device(IP, acceptLanguage: al);
+            var device = new Device(acceptLanguage: al);
             Assert.Equal(al, device.AcceptLanguage);
         }
 
         [Fact]
         public void TestSessionAge()
         {
-            var device = new Device(IP, sessionAge: 3600);
+            var device = new Device(sessionAge: 3600);
             Assert.Equal(3600, device.SessionAge);
         }
 
         [Fact]
         public void TestSessionAgeIsNegative()
         {
-            Assert.Throws<ArgumentException>(() => new Device(IP, sessionAge: -1));
+            Assert.Throws<ArgumentException>(() => new Device(sessionAge: -1));
         }
 
         [Fact]
         public void TestSessionId()
         {
-            var device = new Device(IP, sessionId: "foo");
+            var device = new Device(sessionId: "foo");
             Assert.Equal("foo", device.SessionId);
         }
 
         [Fact]
         public void TestSessionIdIsTooLong()
         {
-            Assert.Throws<ArgumentException>(() => new Device(IP, sessionId: new string('x', 256)));
+            Assert.Throws<ArgumentException>(() => new Device(sessionId: new string('x', 256)));
         }
     }
 }