+++ MaxMind.MinFraud/Request/Device.cs
@@ -21,13 +21,13 @@ namespace MaxMind.MinFraud.Request
         /// the device used in the transaction.</param>
         /// <param name="sessionAge">The number of seconds between the
         /// creation of the user's session and the time of the transaction.
-        /// Note that sessionAge is not the duration of the current visit, but 
+        /// Note that sessionAge is not the duration of the current visit, but
         /// the time since the start of the first visit.</param>
         /// <param name="sessionId">A string up to 255 characters in length.
         /// This is an ID that uniquely identifies a visitor's session on the
         /// site.</param>
         public Device(
-            IPAddress ipAddress,
+            IPAddress? ipAddress = null,
             string? userAgent = null,
             string? acceptLanguage = null,
             double? sessionAge = null,
@@ -56,7 +56,7 @@ namespace MaxMind.MinFraud.Request
         /// in the transaction.
         /// </summary>
         [JsonProperty("ip_address")]
-        public IPAddress IPAddress { get; }
+        public IPAddress? IPAddress { get; }
 
         /// <summary>
         /// The HTTP “User-Agent” header of the browser used in the